% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available.R
\name{available}
\alias{available}
\alias{available_icons}
\alias{available_types}
\title{List available icons or types}
\usage{
available_icons(pattern = NULL, ...)

available_types(pattern = NULL, ...)
}
\arguments{
\item{pattern}{A regular expression pattern}

\item{...}{Additional arguments that are passed to \code{\link[=grepl]{grepl()}}.}
}
\value{
\itemize{
\item \code{available_icons()} returns a named (by type) list of icons.
}

\itemize{
\item \code{available_types()} returns a vector of icon types matching the given pattern.
}
}
\description{
Helper functions for providing a list of available icons or types.
Both functions support the use of regular expressions to subset the results to make
them more manageable.
}
\examples{
available_icons("vcs")


available_types("Common")

}
