% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icon.R
\name{icon}
\alias{icon}
\title{Create an icon}
\usage{
icon(
  name,
  height = NULL,
  width = NULL,
  resize = FALSE,
  output = c("img", "tag")
)
}
\arguments{
\item{name}{Icon name.}

\item{height}{Icon height in pixels.}

\item{width}{Icon width in pixels.}

\item{resize}{Should the icon be resized to the provided height and/or width.}

\item{output}{Return either a "magick-image" object or img tag text string.}
}
\value{
The return type is determined by the value of \code{output} selected.
}
\description{
Creates an icon using based on the provided \code{name}. The icon will be returned
either as a "magick-image" object (if \code{output = "img"}) or as a text string containing an
html img tag (if \code{output = "tag"}).

The height and the width can be specified, the function will try to match the icon with the
closest possible size and can resize if necessary.
}
\examples{

icon("rstudio", 64)
icon("rstudio", 72)
icon("rstudio", 72, resize = TRUE)

icon("rstudio", 64, output = "tag")

}
