% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icon.R
\name{icon_info}
\alias{icon_info}
\title{Icon details}
\usage{
icon_info(name)
}
\arguments{
\item{name}{Icon name.}
}
\value{
A named list.
}
\description{
Retrieves basic useful information about the requested icon, including its type,
available sizes, and image formats.
}
\examples{

icon_info("rstudio")

}
