% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icons.R
\docType{data}
\name{icons}
\alias{icons}
\title{Icons used by RStudio IDE}
\format{
A tibble with 544 rows and 7 columns.
\itemize{
\item name: Icon name
\item type: Icon category
\item data: Raw vectors containing the icon's image data
\item format: Icon image format
\item width: Icon's width in pixels
\item height: Icon's height in pixels
\item filesize: Icon's filesize in bytes
}
}
\source{
\url{https://github.com/rstudio/rstudio/}
}
\usage{
icons
}
\description{
A tibble containing the commonly used icons in the RStudio IDE interface.
Multiple copies of most icons are available at different sizes, and icons
are categorized by their usage within the IDE (\code{type}).
}
\keyword{datasets}
