% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preview.R
\name{preview}
\alias{preview}
\alias{preview_icon}
\alias{preview_type}
\alias{preview_all}
\title{Preview icons}
\usage{
preview_icon(name, max_size = 256, display = TRUE, inc_label = FALSE)

preview_type(type, max_size = 128, display = TRUE)

preview_all(n_col = 4, max_size = 128, display = TRUE)
}
\arguments{
\item{name}{Icon name to preview.}

\item{max_size}{Maximum size of icon to include in preview, checks height and width.}

\item{display}{Should the preview be displayed, via \code{print()} method of "magick-image".}

\item{inc_label}{Should the icons name be include in the preview.}

\item{type}{Icon type to preview.}

\item{n_col}{Number of columns to use when displaying all icons.}
}
\value{
Returns a "magick-image" object.
}
\description{
Helper functions for previewing specific icons, icon types, or all available icons.
}
\examples{

preview_icon("rstudio")

preview_type("File")

}
